update UI: Shop - Bundle tab
