+++ src/components/ui/CardBarista.tsx
@@ -57,7 +57,7 @@ const CardBarista = ({
             {totalBonus}
           </div>
           <div className="w-4 h-4">
-            <img src="/images/bonus.png" alt="" />
+            <img src="/images/kbuck.png" alt="" />
           </div>
         </div>
         {isBarista && (
