+++ src/types/bundle.ts
@@ -1,6 +1,13 @@
 export type Bundle = {
-    id: string;
-    name: string;
-    imageUrl: string;
-    price: number;
+  id: string;
+  name: string;
+  imageUrl: string;
+  price: string;
+  item: [
+    {
+      name: string;
+      value: string;
+      imageUrl: string;
+    }
+  ];
 };
